@@ -12,7 +12,6 @@ import inspect
 import os
 import sys
 from pathlib import Path
-from typing import List, Tuple
 
 # Import plugin update library from maintainers/scripts/pluginupdate.py
 ROOT = Path(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))))  # type: ignore
@@ -21,13 +20,11 @@ sys.path.insert(
 )
 import pluginupdate
 
-GET_PLUGINS = f"""(
-with import <localpkgs> {{ }};
+GET_PLUGINS = f"""with import <localpkgs> {{ }};
 let
   inherit (kakouneUtils.override {{ }}) buildKakounePluginFrom2Nix;
-  generated = callPackage {ROOT}/generated.nix {{
-    inherit buildKakounePluginFrom2Nix;
-  }};
+  generated = callPackage {ROOT}/generated.nix {{ inherit buildKakounePluginFrom2Nix; }};
+
   hasChecksum =
     value:
     lib.isAttrs value
@@ -35,20 +32,23 @@ let
       "src"
       "outputHash"
     ] value;
-  getChecksum =
-    name: value:
-    if hasChecksum value then
-      {{
-        submodules = value.src.fetchSubmodules or false;
-        sha256 = value.src.outputHash;
-        rev = value.src.rev;
-      }}
-    else
-      null;
-  checksums = lib.mapAttrs getChecksum generated;
+
+  parse = name: value: {{
+    pname = value.pname;
+    version = value.version;
+    homePage = value.meta.homepage;
+    checksum =
+      if hasChecksum value then
+        {{
+          submodules = value.src.fetchSubmodules or false;
+          sha256 = value.src.outputHash;
+          rev = value.src.rev;
+        }}
+      else
+        null;
+  }};
 in
-lib.filterAttrs (n: v: v != null) checksums
-)"""
+lib.mapAttrs parse generated"""
 
 HEADER = "# This file has been @generated by ./pkgs/applications/editors/kakoune/plugins/update.py. Do not edit!"
 
@@ -56,7 +56,7 @@ HEADER = "# This file has been @generated by ./pkgs/applications/editors/kakoune
 class KakouneEditor(pluginupdate.Editor):
     def generate_nix(
         self,
-        plugins: List[Tuple[pluginupdate.PluginDesc, pluginupdate.Plugin]],
+        plugins: list[tuple[pluginupdate.PluginDesc, pluginupdate.Plugin]],
         outfile: str,
     ):
         with open(outfile, "w+") as f:
